requirements.txt
python-irodsclient==3.0.0
